Move cached parser file into a separate directory
