+++ chuffed/flatzinc/lexer.lxx
@@ -32,7 +32,7 @@ void yyerror(void*, const char*);
 #define yyerror(s) yyerror(yyextra, s)
 
 #include <chuffed/flatzinc/flatzinc.h>
-#include <chuffed/flatzinc/parser.tab.h>
+#include <chuffed/flatzinc/generated_parser/parser.tab.h>
 
 const char* stringbuf;
 int stringbuflen;
